### Before

```
$ time bundle exec ruby bin/github-nippou -s 20160325
(snip)
bundle exec ruby bin/github-nippou -s 20160325  1.50s user 0.27s system 10% cpu 16.724 total
```

### After

```
$ time bundle exec ruby bin/github-nippou -s 20160325
(snip)
bundle exec ruby bin/github-nippou -s 20160325  1.49s user 0.28s system 16% cpu 10.603 total
```